Refine selectFunction fallback
